Rename pkg runner to contextx
