+++ README.md
@@ -3,15 +3,16 @@
 ## Packages
 
 There are two types of packages: 
-- `ext`: extensions to the stdlib, they augment stdlib package functionalities. Usually named after a stdlib package name plus `x`.
+- `ext`: extensions to the stdlib, they augment stdlib package functionalities.
+  Usually named after a stdlib package name plus `x`.
 - `new`: they provide logic not present in the stdlib.
 
-| Package            | Type  | Description                                                         |
-| ---                | ----  | -----------                                                         |
-| [`httpx`](httpx)   | `ext` | Extends package `net/http` with routing utilities, decorators, etc  |
-| [`logx`](logx)     | `ext` | Adds structured logging and logstash support to `log`               |
-| [`runner`](runner) | `new` | Provides types and decorators to run tasks with a `context.Context` |
-| [`timex`](timex)   | `ext` | Extends `time` with other time functions and utilities              |
+| Package                | Type  | Description                                                                       |
+| ---                    | ----  | -----------                                                                       |
+| [`httpx`](httpx)       | `ext` | Extends `net/http` with routing utilities, decorators, etc                        |
+| [`logx`](logx)         | `ext` | Extends `log` with structured logging and logstash support                        |
+| [`contextx`](contextx) | `ext` | Extends `context` with types and decorators to run tasks with a `context.Context` |
+| [`timex`](timex)       | `ext` | Extends `time` with other time functions and utilities                            |
 
 ## Contributors
