+++ contextx/posixsignal_test.go
@@ -1,4 +1,4 @@
-package runner
+package contextx
 
 import (
 	"context"
